requirements.txt
heyoo==0.1.2
numpy==1.26.4
openai==1.58.1
pandas==1.5.1
Pillow==10.4.0
requests==2.28.1
streamlit==1.41.1
gspread==5.6.2
oauth2client==4.1.3
urllib3==1.26.12
protobuf==4.25.2
PyOpenSSL==24.3.0
pytz==2024.2
aiohttp==3.9.1
secure-smtplib==0.1.1
asyncio==3.4.3

 

